moved all relevant files into the request activation key of their subdirectory
