+++ service-front/app/src/Actor/src/Form/RequestActivationKey/CreateNewActivationKey.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Actor\Form;
+namespace Actor\Form\RequestActivationKey;
 
 use Common\Form\AbstractForm;
 use Mezzio\Csrf\CsrfGuardInterface;
