+++ service-front/app/src/Actor/src/Form/RequestActivationKey/RequestContactDetails.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Actor\Form;
+namespace Actor\Form\RequestActivationKey;
 
 use Actor\Validator\OptionSelectedValidator;
 use Common\Form\AbstractForm;
